Staging to master (#1830)
* fixing https requests for year/rais and year/secex (#1809)

* add new endpoint to get years in datasets (#1810)

* Feature/fixing bar request (#1811)

* Removing console.log

* Making the python file do the problematic request

* add request responde in javascript

* remove logs

---------


* Feature/updating ranking (#1812)

* Renaming ranking tabs

* Changing the economic activities ranking

* Changing location and  product rankings data and the layout of ranking page

* removing console .log

* Adding new atributes to secex_Ymb and rais_Yi tables and implementing them on the ranking screens

---------


* change years database endpoint (#1813)

* Feature/update our team (#1815)

* new members

* new member

* adding space for new members

* adding new members and links to the about page

---------


* updating urls to support the https (#1816)

* Feature/changingEvolutionGraph (#1818)

* adding years

* divide requests per year

* Getting the years dynamically

---------


* Feature/about update (#1817)

* fernanda github link fix

* remove ex-member

* add institutions names

* fixing responsivity

---------


* Update views.py (#1819)

* Feature/new blog (#1821)

* Changing blog main layout

* Finishing filter's section and adding more styles

* Fixing parameter

* Trying to fix blog's search bar

* Fixing blog search bar

---------


* Feature/new news page (#1822)

* updating news page (layout and filter)

* updating layout and translations for News page

* updating search filter to include author and text content in News page

* Fixing news image display and filtering error

---------


* Feature/studies section (#1823)

* Changing blog main layout

* add new text and style to studies section

* Finishing filter's section and adding more styles

* Fixing parameter

* add new styles and base html in studiesSection

* add filter by keyword and some style features

* add translate for date and view features

* add search function and mock image to layout

* add image form and data

* remove comment

* add new scholar preview in show page and default image to studies

* add new search filters to blond and scholar

* Adding the article image to the admin edit page and to the admin's show page;
Adding translations that were missing

* Adding redirect functions of scholar section to a separated file and fixing it's search bar

* Adding condition to search an article by author's name

* Adding borderBottom on blog and news sidebar

---------


* Fixing scholar translation

* Feature/blog&newsResponsivity (#1824)

* Fixing blog and news responsivity

* Adding responsiveness to the study page

---------


* Feature/update new members (#1825)

* adding templates for new members

* updating members in About page

* fixing translations in About page

* feature/scholarAdjustments (#1826)

* fix scholar form and edit about page

* Add new image styles for blog image and change crop definition in form

* Feature/newScholarDateField (#1827)

* add new datetime field to scholar

* Creating plucation data field on article registration page

* Fixing scholar form problem with publication_data field

---------


* Changing font-size and font-family of blog posts

* SQLAlchemy version change for compatibility with SQL version

* Change SQLAlchemy version

* hf

* updating mysql-python version

---------

+++ dataviva/api/attrs/models.py
@@ -138,7 +138,7 @@ class School(db.Model, AutoSerialize, ExpandedAttr, ImageAttr):
 
     __tablename__ = 'attrs_school'
     id = db.Column(db.String(8), primary_key=True)
-    is_vocational = db.Column(db.Integer(1))
+    is_vocational = db.Column(db.Integer())
     school_type_id = db.Column(db.String(1))
     school_type_en = db.Column(db.String(32))
     school_type_pt = db.Column(db.String(32))
@@ -198,8 +198,8 @@ class Wld(db.Model, AutoSerialize, BasicAttr, ImageAttr):
     id = db.Column(db.String(5), primary_key=True)
     id_2char = db.Column(db.String(2))
     id_3char = db.Column(db.String(3))
-    id_num = db.Column(db.Integer(11))
-    id_mdic = db.Column(db.Integer(11))
+    id_num = db.Column(db.Integer())
+    id_mdic = db.Column(db.Integer())
 
     ymw = db.relationship("Ymw", backref = 'wld', lazy = 'dynamic')
     ympw = db.relationship("Ympw", backref = 'wld', lazy = 'dynamic')
@@ -231,7 +231,7 @@ class Bra(db.Model, AutoSerialize, BasicAttr, ImageAttr):
 
     __tablename__ = 'attrs_bra'
     id = db.Column(db.String(10), primary_key=True)
-    id_ibge = db.Column(db.Integer(7))
+    id_ibge = db.Column(db.Integer())
     abbreviation = db.Column(db.String(2))
 
     distance = 0
@@ -322,7 +322,7 @@ class Distances(db.Model):
 class Yb(db.Model, AutoSerialize):
 
     __tablename__ = 'attrs_yb'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     bra_id = db.Column(db.String(10), db.ForeignKey(Bra.id), primary_key=True)
     population = db.Column(db.Integer)
 
@@ -353,7 +353,7 @@ class Bs(db.Model, AutoSerialize):
 class Ybs(db.Model, AutoSerialize):
 
     __tablename__ = 'attrs_ybs'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     bra_id = db.Column(db.String(10), db.ForeignKey(Bra.id), primary_key=True)
     stat_id = db.Column(db.String(20), db.ForeignKey(Stat.id), primary_key=True)
     stat_val = db.Column(db.Float)
@@ -364,7 +364,7 @@ class Ybs(db.Model, AutoSerialize):
 class Ybb(db.Model, AutoSerialize):
 
     __tablename__ = 'attrs_ybb'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     bra_id = db.Column(db.String(10), db.ForeignKey(Bra.id), primary_key=True)
     bra_id_target = db.Column(db.String(10), db.ForeignKey(Bra.id), primary_key=True)
     prox_cnae = db.Column(db.Float)
@@ -378,7 +378,7 @@ class Ybb(db.Model, AutoSerialize):
 class Ypp(db.Model, AutoSerialize):
 
     __tablename__ = 'attrs_ypp'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     hs_id = db.Column(db.String(8), db.ForeignKey(Hs.id), primary_key=True)
     hs_id_target = db.Column(db.String(8), db.ForeignKey(Hs.id), primary_key=True)
     prox_bra = db.Column(db.Float)
@@ -389,7 +389,7 @@ class Ypp(db.Model, AutoSerialize):
 class Yww(db.Model, AutoSerialize):
 
     __tablename__ = 'attrs_yww'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     wld_id = db.Column(db.String(5), db.ForeignKey(Wld.id), primary_key=True)
     wld_id_target = db.Column(db.String(5), db.ForeignKey(Wld.id), primary_key=True)
     prox_bra = db.Column(db.Float)
@@ -400,7 +400,7 @@ class Yww(db.Model, AutoSerialize):
 class Yii(db.Model, AutoSerialize):
 
     __tablename__ = 'attrs_yii'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     cnae_id = db.Column(db.String(8), db.ForeignKey(Cnae.id), primary_key=True)
     cnae_id_target = db.Column(db.String(8), db.ForeignKey(Cnae.id), primary_key=True)
     prox_bra = db.Column(db.Float)
@@ -411,7 +411,7 @@ class Yii(db.Model, AutoSerialize):
 class Yoo(db.Model, AutoSerialize):
 
     __tablename__ = 'attrs_yoo'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     cbo_id = db.Column(db.String(6), db.ForeignKey(Cbo.id), primary_key=True)
     cbo_id_target = db.Column(db.String(6), db.ForeignKey(Cbo.id), primary_key=True)
     prox_bra = db.Column(db.Float)
@@ -422,7 +422,7 @@ class Yoo(db.Model, AutoSerialize):
 class Yuu(db.Model, AutoSerialize):
 
     __tablename__ = 'attrs_yuu'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     university_id = db.Column(db.String(6), db.ForeignKey(Cbo.id), primary_key=True)
     university_id_target = db.Column(db.String(6), db.ForeignKey(Cbo.id), primary_key=True)
     prox_course_hedu = db.Column(db.Float)
